+++ src/Shakebook/Pandoc.hs
@@ -33,19 +33,17 @@ module Shakebook.Pandoc (
 , prefixAllImages
 , progressivelyDemoteHeaders
 , replaceUnusableImages
-
--- * Lenses
-, viewContent
-, viewSrcPath
-, viewUrl
 ) where
 
+import Composite.Aeson
+import Composite.Record
 import           Control.Comonad
 import           Control.Comonad.Cofree
 import           Data.Aeson
+import Data.Aeson.BetterErrors
 import           Data.Aeson.Lens
 import           Data.Aeson.With
-import           Development.Shake.Plus
+import           Development.Shake.Plus hiding ((:->))
 import           Path.Extensions
 import           RIO
 import qualified RIO.ByteString.Lazy    as LBS
@@ -138,49 +136,19 @@ prefixAllImages dir = walk handleImages where
 flattenMeta :: MonadAction m => (Pandoc -> PandocIO Text) -> Meta -> m Value
 flattenMeta opts meta = liftAction $ Slick.Pandoc.flattenMeta opts meta
 
--- | View the "content" field of a JSON value.
-viewContent :: ToJSON a => a -> Text
-viewContent = view' (key "content" . _String)
-
--- | Add "content" field from input Text.
-withContent :: Text -> Value -> Value
-withContent = withStringField "content"
-
--- | View the "src-path" field of a JSON Value.
-viewSrcPath :: ToJSON a => a -> Text
-viewSrcPath = view' (key "src-path" . _String)
-
--- | Add "src-path" field based on input Text.
-withSrcPath :: Text -> Value -> Value
-withSrcPath = withStringField "src-path"
-
--- | View the "url" field of a JSON Value.
-viewUrl :: ToJSON a => a -> Text
-viewUrl = view' (key "url" . _String)
-
--- | Add "url" field from input Text.
-withUrl :: Text -> Value -> Value
-withUrl = withStringField "url"
-
--- | Add a leading slash to a `Path Rel File` to turn it into a url as `Text`.
-toGroundedUrl :: Path Rel File -> Text
-toGroundedUrl = T.pack . toFilePath . ($(mkAbsDir "/") </>)
-
 {-|
   Get a JSON Value of Markdown Data with markdown body as "contents" field
   and the srcPath as "srcPath" field.
 -}
-loadMarkdownAsJSON :: (MonadAction m, MonadThrow m)
+loadMarkdownAsJSON :: (MonadAction m, MonadThrow m, FileLike Rel a)
                    => ReaderOptions
                    -> WriterOptions
-                   -> Within Rel (Path Rel File)
+                   -> a
                    -> m Value
 loadMarkdownAsJSON ropts wopts srcPath = do
   pdoc@(Pandoc meta _) <- readMarkdownFile ropts srcPath
   meta' <- flattenMeta (writeHtml5String wopts) meta
   outText <- runPandocA $ writeHtml5String wopts pdoc
-  supposedUrl <- toGroundedUrl <$> withHtmlExtension (extract srcPath)
-  return $ withContent outText
-         . withSrcPath (T.pack . toFilePath $ extract srcPath)
-         . withUrl supposedUrl $ meta'
-
+  return $ withStringField "src-path" (T.pack $ toFilePath $ toFile srcPath)
+         $ withStringField "content" outText
+         $ meta'